Add IdSource base class
Common base for different sources for identifying a user. IdSource:s will
be grouped together in a IdSource::Group, which is the main class through
which the rest of the daemon will interface with sources.

Just instantiate a IdSource::Group in Daemon for now. Will actually be used
in following commit that implements the D-Bus interface.

Signed-off-by: Martin Ejdestig <mejdestig@luxoft.com>