Make it possible to set sources to enable in a configuration file
So that a single binary can be used for different products.

The format used is "key file"/"ini file" since there is a parser readily
available in Glib for this (also used by other daemons that uses Glib, e.g.
NetworkManager). See
https://developer.gnome.org/glib/stable/glib-Key-value-file-parser.html for
details.

ScopedTempFile, that creates a temporary file in /tmp, is added to actually
use real files in unit tests instead of adding a function for e.g. parsing
from a string (Glib has API for that but better to test real code path).

Signed-off-by: Martin Ejdestig <mejdestig@luxoft.com>