fixup! Add mass storage device ID source
Add tests for parsing of mass storage device ID source file. Did not have
ScopedTempFile when first implementing the source. Originally looked into
writing tests by mocking Gio::VolumeMonitor but it is... more complicated
than it is worth. If we want to test the MSD source more we should probably
do it on target and create real mount points.

Not entierly happy with having MassStorageDeviceIdSource::Parser but it
separates it a bit from the actual source.

Also rename read_file() to read_file_and_notify() to better describe what
it does... hopefully.