fixup! Add IdSource base class
Add max identified users to config.h. Will use it in src/cli/arguments.cpp
description for --identified-user option in following commit.

Keep the constant in IdSource::Group and just set value from config.h.
Makes it so "config.h" does not have to be included everywhere were value
is used.

Not entirely happy with having it in config.h at all, in fact. But I...
guess it makes... a bit of sense... if we think of the value as something
we might want to... configure... meh.