Add smart card identification source
A PCSCContext singleton is added for isolating that all SCard API calls are
made in a separate thread and for allow to share the connection to the
pscs-lite daemon between different sources at some later point (e.g. if an
NFC source is added that uses pcsc-lite).

Currently only extracts UID from contactless smart cards and uses it as a
way to identify a user.

In the future it might make sense to map different readers to different
seats. The name of the reader is therefor passed along when an UID is
extracted. No mapping between reader and seat id is done yet though.

An IdleQueue helper class is added for pushing information from the thread
in PCSCContext to the main thread.

Signed-off-by: Martin Ejdestig <mejdestig@luxoft.com>