Configure clang-tidy so braces are always added for do/for/if/while
We have said "strive to follow AUTOSAR C++14" and rule M6-3-1 and M6-4-1
says "The statement forming the body of a switch, while, do ... statement
shall be a compound statement." and "An if ( condition ) construct shall
be followed by a compound statement. The else keyword shall be followed by
either a compound statement, or another if statement."

It adds a little bit of clutter and there are other checks for checking
things that these rules are meant to prevent (e.g. wrongly written
multi-line macros are caught by bugprone-multiple-statement-macro). But
since it is mostly harmless and we have a way to enforce it with
clang-tidy, do so.

Signed-off-by: Martin Ejdestig <mejdestig@luxoft.com>